+++ lib/src/components/Modal/components/Assets/assets.module.scss
@@ -1,0 +1,49 @@
+@layer components {
+  .assets {
+    --contentWidthMobile: calc(100vw - 2 * 1rem);
+    --contentHeight: calc(100vh - 2 * 5rem);
+    --contentWidthDesktop: calc(100vw - 2 * 2rem);
+  }
+
+  @mixin sizeOfElements {
+    width: var(--contentWidthMobile);
+    max-height: var(--contentHeight);
+
+    @media (min-width: md) {
+      width: var(--contentWidthDesktop);
+    }
+
+    /* for big screens */
+    @media (min-width: 1700px) {
+      max-width: 1600;
+    }
+  }
+  .close-button {
+    position: absolute;
+    right: var(--spacing-xl);
+    top: var(--spacing-xl);
+
+    /* Hack for secondary variant on dark mode */
+    color: var(--color-neutral-90);
+    border-color: var(--color-neutral-10);
+    background-color: var(--color-neutral-10);
+
+    &:hover {
+      background-color: var(--color-neutral-40);
+    }
+  }
+
+  .iframe {
+    @include sizeOfElements;
+    flex: 1;
+
+    @media (min-width: md) {
+      aspect-ratio: 16 / 9;
+    }
+
+    iframe {
+      width: 100%;
+      height: 100%;
+    }
+  }
+}
