+++ lib/src/components/Modal/components/Assets/types.ts
@@ -1,0 +1,14 @@
+import type { DialogOptions } from '@ariakit/react'
+
+export type AssetWithTitleProps = {
+  children: React.ReactNode
+  customContent?: JSX.Element | string
+  subtitle?: JSX.Element | string
+  title?: JSX.Element | string
+}
+
+export type BackdropProps = Pick<DialogOptions, 'hideOnInteractOutside'>
+
+export type IframeProps = {
+  children: JSX.Element
+}
