+++ lib/src/components/Modal/components/Content/types.ts
@@ -1,0 +1,13 @@
+import type { UseModal } from '@/components/Modal/types'
+
+export interface ContentOptions {
+  children: React.ReactNode
+}
+
+export type ContentProps = ContentOptions & {
+  store: UseModal
+  /**
+   * show or hide the closing button
+   */
+  withClosingButton?: boolean
+}
