+++ lib/src/components/Modal/components/Header/header.module.scss
@@ -1,0 +1,23 @@
+@layer components {
+  .header {
+    position: sticky;
+    top: 0;
+    z-index: 1;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: var(--color-neutral-10);
+    border-top-left-radius: var(--radius-md);
+    border-top-right-radius: var(--radius-md);
+    padding-bottom: var(--spacing-xxl);
+    padding-left: var(--spacing-xxl);
+    padding-right: var(--spacing-3xl);
+    padding-top: var(--spacing-xxl);
+  }
+
+  .header-subtitle {
+    color: var(--color-neutral-70);
+    margin: 0;
+  }
+}
