+++ lib/src/components/Modal/components/Header/types.ts
@@ -1,0 +1,11 @@
+import type { HTMLAttributes, ReactElement } from 'react'
+
+import type { MergeProps } from '@/utils/forwardRefWithAs'
+
+export interface HeaderOptions {
+  icon?: ReactElement
+  subtitle?: JSX.Element | string
+  title: JSX.Element | string
+}
+
+export type HeaderProps = MergeProps<HeaderOptions, HTMLAttributes<HTMLDivElement>>
