+++ lib/src/components/Modal/docs/examples/sizes.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
-import type { ModalProps } from '@/components/Modal'
 import { Modal, useModal } from '@/components/Modal'
+import type { ModalProps } from '@/components/Modal/types'
 import { Button } from '@old/Button'
 
 const Example = () => {
