+++ lib/src/components/Modal/docs/properties.json
@@ -9,12 +9,12 @@
         "description": "",
         "name": "ariaLabel",
         "parent": {
-          "fileName": "welcome-ui/lib/src/components/Modal/index.tsx",
+          "fileName": "welcome-ui/lib/src/components/Modal/types.ts",
           "name": "ModalOptions"
         },
         "declarations": [
           {
-            "fileName": "welcome-ui/lib/src/components/Modal/index.tsx",
+            "fileName": "welcome-ui/lib/src/components/Modal/types.ts",
             "name": "ModalOptions"
           }
         ],
@@ -28,12 +28,12 @@
         "description": "",
         "name": "size",
         "parent": {
-          "fileName": "welcome-ui/lib/src/components/Modal/index.tsx",
+          "fileName": "welcome-ui/lib/src/components/Modal/types.ts",
           "name": "ModalOptions"
         },
         "declarations": [
           {
-            "fileName": "welcome-ui/lib/src/components/Modal/index.tsx",
+            "fileName": "welcome-ui/lib/src/components/Modal/types.ts",
             "name": "ModalOptions"
           }
         ],
