+++ lib/src/components/Modal/index.tsx
@@ -1,47 +1,33 @@
-import * as Ariakit from '@ariakit/react'
+import { Dialog, DialogDisclosure, useDialogStore } from '@ariakit/react'
 import { useTheme } from '@xstyled/styled-components'
-import type { HTMLAttributes, PropsWithChildren } from 'react'
 import React, { cloneElement, forwardRef } from 'react'
 
 import { classNames } from '@/utils'
-import type { MergeProps } from '@/utils/forwardRefWithAs'
 import { forwardRefWithAs } from '@/utils/forwardRefWithAs'
 //TODO Migrate shape, but to what?
 import type { ShapeProps } from '@old/Shape'
 import { Shape } from '@old/Shape'
 
-import { Assets } from './Assets'
-import { Content } from './Content'
-import { Footer } from './Footer'
-import { Header } from './Header'
+import { Assets } from './components/Assets'
+import { Content } from './components/Content'
+import { Footer } from './components/Footer'
+import { Header } from './components/Header'
 import modalStyles from './modal.module.scss'
-import type { Size } from './theme'
-
-export interface ModalOptions extends Omit<Ariakit.DialogOptions<'div'>, 'as'> {
-  ariaLabel: string
-  children: React.ReactElement
-  size?: Size
-}
-
-export type ModalProps = MergeProps<ModalOptions, HTMLAttributes<HTMLDivElement>>
-export type UseModal = Ariakit.DialogStore
-export type UseModalProps = Ariakit.DialogStoreProps & {
-  /**
-   * Call a function before closing the modal
-   * @deprecated use onClose on <Modal /> instead
-   */
-  onClose?: () => void
-}
-export type UseModalState = Ariakit.DialogStoreState
-
-type BackdropProps = Pick<ModalOptions, 'backdrop' | 'hideOnInteractOutside'>
+import type {
+  BackdropProps,
+  BodyProps,
+  ModalProps,
+  TriggerProps,
+  UseModal,
+  UseModalProps,
+} from './types'
 
 const cx = classNames(modalStyles)
 
 export function useModal(options?: UseModalProps): UseModal {
   const { onClose, setOpen, ...storeOptions } = options || {}
 
-  const dialog = Ariakit.useDialogStore({
+  const dialog = useDialogStore({
     setOpen: open => {
       if (!open && onClose) {
         onClose()
@@ -87,24 +73,22 @@ const ModalComponent = forwardRefWithAs<ModalProps, 'div'>(
     ref
   ) => {
     return (
-      <Ariakit.Dialog
+      <Dialog
         aria-label={ariaLabel}
         backdrop={<Backdrop backdrop={backdrop} hideOnInteractOutside={hideOnInteractOutside} />}
         hideOnInteractOutside={hideOnInteractOutside}
         ref={ref}
         render={props =>
-          As ? <As {...props} /> : <div className={cx('dialog', `size-${size}`)} {...props} />
+          As ? <As {...props} /> : <div className={cx('root', `size-${size}`)} {...props} />
         }
         store={store}
-        {...(rest as Ariakit.DialogProps<'div'>)}
+        {...rest}
       >
         {children}
-      </Ariakit.Dialog>
+      </Dialog>
     )
   }
 )
-
-type BodyProps = PropsWithChildren<HTMLAttributes<HTMLElement>>
 
 const Body = forwardRef<HTMLElement, BodyProps>((props, ref) => {
   return <section className={cx('body')} ref={ref} {...props} />
@@ -122,11 +106,9 @@ const Cover: React.FC<ShapeProps> = props => {
   )
 }
 
-type TriggerProps = PropsWithChildren<{ store: Ariakit.DialogStore }>
-
 const Trigger = forwardRefWithAs<TriggerProps, 'button'>(({ as: As, store, ...rest }, ref) => {
   return (
-    <Ariakit.DialogDisclosure
+    <DialogDisclosure
       ref={ref}
       render={As ? props => <As {...props} /> : undefined}
       store={store}
