+++ lib/src/components/Modal/modal.module.scss
@@ -19,7 +19,7 @@
     }
   }
 
-  .dialog {
+  .root {
     background-color: var(--color-neutral-10);
     border-color: var(--color-neutral-30);
     border-style: solid;
@@ -87,56 +87,4 @@
     padding-right: var(--spacing-3xl);
     padding-bottom: var(--spacing-xxl);
   }
-
-  .header {
-    position: sticky;
-    top: 0;
-    z-index: 1;
-    width: 100%;
-    /* flex-shrink does nothing, the parent is not a flex */
-    flex-shrink: 0;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: var(--color-neutral-10);
-    border-top-left-radius: var(--radius-md);
-    border-top-right-radius: var(--radius-md);
-    padding-bottom: var(--spacing-xxl);
-    padding-left: var(--spacing-xxl);
-    padding-right: var(--spacing-3xl);
-    padding-top: var(--spacing-xxl);
-  }
-
-  .header-subtitle {
-    color: var(--color-neutral-70);
-    margin: 0;
-    /* margin-bottom: 0; TESTME */
-    /* variant: 'sm', TESTME */
-  }
-
-  .footer {
-    position: sticky;
-    bottom: 0;
-    width: 100%;
-    flex-shrink: 0;
-    z-index: 1;
-    background-color: var(--color-neutral-10);
-    border-bottom-left-radius: var(--radius-md);
-    border-bottom-right-radius: var(--radius-md);
-    border-top: 0 solid var(--color-neutral-30);
-  }
-
-  .footer-children-wrapper {
-    padding-block: var(--spacing-xl);
-    padding-inline: var(--spacing-xxl);
-  }
-
-  .footer-information {
-    background-color: var(--color-beige-20);
-    padding-bottom: var(--spacing-xxl);
-    padding-left: var(--spacing-xxl);
-    padding-right: var(--spacing-xxl);
-    padding-top: var(--spacing-xl);
-  }
-  /* end of layer */
 }
